Moved Platoon-relevant Services to own folder
