+++ lib/PlatoonService/src/ProcessingChainFactory.cpp
