+++ platformio.ini
@@ -108,6 +108,7 @@ lib_deps =
     Utilities
     gabryelreyes/SerialMuxProt@^2.0.0
     bblanchon/ArduinoJson@^6.21.3
+    PlatoonService
 lib_ignore =
     ConvoyFollower
     RemoteControl
